Implement Quarkus

+++ test-crud-quarkus/src/main/java/com/waizly/service/JurusanService.java
@@ -4,22 +4,29 @@ import java.util.List;
 import java.util.Optional;
 
 import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 import javax.ws.rs.core.Response;
 
+import com.waizly.dto.ResponseData;
 import com.waizly.models.Jurusan;
 
 @ApplicationScoped
 @Transactional
 public class JurusanService {
     
+
+    @Inject
+    EntityManager entitas;
+
     public List<Jurusan> getAll() {
         return Jurusan.listAll();
     }
 
     public Response findByjurusanName(String name) {
-        Iterable<Jurusan> jurusan = Jurusan.find(
+        List<Jurusan> jurusan = Jurusan.find(
         "lower(nameJurusan) LIKE CONCAT ('%', lower(?1), '%')", name).list();
         if(jurusan ==  null) {
             return Response.status(Response.Status.BAD_REQUEST).entity("JURUSAN_NOT_FOUND").build();
@@ -27,29 +34,42 @@ public class JurusanService {
         return Response.ok(jurusan).build();
     }
 
-    public Jurusan create(@Valid Jurusan jurusan) {
+    public Response create(@Valid Jurusan jurusan) {
+        ResponseData responData = new ResponseData();
+        if(jurusan == null) {
+            responData.setMessages("Data gagal");
+            return Response.status(401).entity(responData).build();
+        }
         jurusan.persist();
-        return jurusan;
+        responData.setStatus(true);
+        responData.setMessages("Berhasil");
+        responData.setData(jurusan);
+        return Response.ok(responData).build();
     }
 
     public Response update(String kodeJurusan, @Valid Jurusan newJurusan) {
+        ResponseData responData = new ResponseData();
         Optional<Jurusan> optionalJurusan = Jurusan.find(
         "kodeJurusan = ?1", kodeJurusan).firstResultOptional();
         if (optionalJurusan.isEmpty()) {
-            return Response.status(Response.Status.BAD_REQUEST).entity("KODEJURUSAN_NOT_FOUND").build();
+            responData.setMessages("Fales");
+            return Response.status(400).entity("KODEJURUSAN_NOT_FOUND").build();
         }
         Jurusan oldJurusan = optionalJurusan.get();
         oldJurusan.setNameJurusan(newJurusan.getNameJurusan());
         oldJurusan.setJenjang(newJurusan.getJenjang());
-        return Response.ok(oldJurusan).build();
+        responData.setStatus(true);
+        responData.setMessages("Berhasil");
+        responData.setData(oldJurusan);
+        return Response.ok(responData).build();
     }
 
     public Response deleteId(Long idJurusan) {
         Jurusan jurusan = Jurusan.findById(idJurusan);
-        Boolean deleteJurusan = Jurusan.deleteById(idJurusan);
-        if(deleteJurusan) {
-            return Response.ok(jurusan).entity("Jurusan success for delete").build();
+        if(jurusan == null) {
+            return Response.status(400).entity("ID NOT_FOUND").build();
         }
-        return Response.status(Response.Status.BAD_REQUEST).entity("ID NOT_FOUND").build();
+        Jurusan.deleteById(idJurusan);
+        return Response.ok(jurusan).entity("Jurusan success for delete").build();
     }
 }
