+++ test-crud-quarkus/src/main/java/com/waizly/service/MatakuliahService.java
@@ -9,13 +9,11 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import javax.ws.rs.core.Response;
 
-import com.waizly.models.Mahasiswa;
+import com.waizly.dto.ResponseData;
 import com.waizly.models.Matakuliah;
 
-import io.quarkus.panache.common.Sort;
 
-
-@ApplicationScoped
+@ApplicationScoped 
 @Transactional
 public class MatakuliahService {
     
@@ -28,7 +26,7 @@ public class MatakuliahService {
     }
     
     public Response findByMatakuliahName(String nameMatakuliah) {
-        Iterable<Matakuliah> mataKuliah = Matakuliah.find(
+        List<Matakuliah> mataKuliah = Matakuliah.find(
         "lower(nameMatakuliah) like concat ('%', lower(?1), '%')", nameMatakuliah).list();
         if(mataKuliah == null) {
             return Response.status(Response.Status.BAD_REQUEST).entity("MATAKULIAH_NOT_FOUND").build();
@@ -36,37 +34,42 @@ public class MatakuliahService {
         return Response.ok(mataKuliah).build();
     }
 
-    public Matakuliah create(@Valid Matakuliah mataKuliah) {
+    public Response create(@Valid Matakuliah mataKuliah) {
+        ResponseData responData = new ResponseData();
+        if(mataKuliah == null) {
+            responData.setMessages("Gagal");
+            return Response.status(401).entity(responData).build();
+        }
         mataKuliah.persist();
-        return mataKuliah;
-    }
-
-    public List<Matakuliah> addMatakuliah(Long idMatakuliah, Long idMahasiswa) {
-        Matakuliah matakuliah = Matakuliah.findById(idMatakuliah);
-        Mahasiswa mahasiswa = Mahasiswa.findById(idMahasiswa);
-        matakuliah.mahasiswaList.add(mahasiswa);
-        matakuliah.persist();
-        return Matakuliah.listAll(Sort.ascending("id"));
+        responData.setStatus(true);
+        responData.setMessages("Matakuliah berhasil ditambahkan");
+        responData.setData(mataKuliah);
+        return Response.ok(responData).build();
     }
 
     public Response update(Long idMatakuliah, @Valid Matakuliah newMataKuliah) {
+        ResponseData responData = new ResponseData();
         Matakuliah oldMataKuliah = Matakuliah.findById(idMatakuliah);
         if (oldMataKuliah == null) {
-            return Response.status(Response.Status.BAD_REQUEST).entity("ID NOT_FOUND").build();
+            responData.setMessages("False");
+            return Response.status(401).entity(responData).build();
         }
         oldMataKuliah.setNameMatakuliah(newMataKuliah.getNameMatakuliah());
         oldMataKuliah.setJumlahSKS(newMataKuliah.getJumlahSKS());
-        return Response.ok(oldMataKuliah).build();
+        responData.setStatus(true);
+        responData.setMessages("berhasil");
+        responData.setData(oldMataKuliah);
+        return Response.ok(responData).build();
     }
 
     public Response deleteId(Long idMatakuliah) {
         entitas.createNativeQuery("DELETE FROM mahasiswa_kuliah WHERE kuliah_id = :id")
         .setParameter("id", idMatakuliah).executeUpdate();
         Matakuliah mataKuliah = Matakuliah.findById(idMatakuliah);
-        Boolean deleteMatakuliah = Matakuliah.deleteById(idMatakuliah);
-        if(deleteMatakuliah) {
-            return Response.ok(mataKuliah).entity("Matakuliah succes for delete").build();
+        if(!mataKuliah.isPersistent()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity("ID NOT_FOUND").build();
         }
-        return Response.status(Response.Status.BAD_REQUEST).entity("ID NOT_FOUND").build();
+        Matakuliah.deleteById(idMatakuliah);
+        return Response.ok(mataKuliah).entity("Matakuliah succes for delete").build();
     }
 }
